use correct type

+++ src/random.c
@@ -4,6 +4,8 @@
 
 #include "random.h"
 
+#include <stddef.h>
+
 static __thread uint64_t x;      // splitmix state
 static __thread uint64_t s[4];   // xoshiro256** state
 
@@ -50,7 +52,7 @@ void jump(void) {
     uint64_t s1 = 0;
     uint64_t s2 = 0;
     uint64_t s3 = 0;
-    for(uint64_t i=0; i < sizeof JUMP / sizeof *JUMP; i++)
+    for(size_t i=0; i < sizeof JUMP / sizeof *JUMP; i++)
         for(int b=0; b < 64; b++) {
             if (JUMP[i] & UINT64_C(1) << b) {
                 s0 ^= s[0];
